requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
cycler==0.11.0
deflate-dict==1.0.9
dict-hash==1.1.16
discord==1.0.1
discord.py==1.7.3
distro==1.6.0
docopt==0.6.2
Equation==1.2.1
ffmpeg==1.4
google-api-core==1.31.5
google-api-python-client==2.1.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==2.8
kiwisolver==1.3.2
matplotlib==3.4.1
mcrcon==0.6.0
multidict==5.2.0
num2words==0.5.10
numpy==1.20.1
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
prettytable==2.1.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.17.0
pytz==2021.3
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sortedcontainers==2.3.0
tabula-py==2.3.0
tabulate==0.8.9
tinydb==4.4.0
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.2.5
yarl==1.7.2
